+++ JitsiForAndroid/app/src/main/java/com/bu/meet/client/FacebookFragment.java
@@ -47,6 +47,8 @@ public class FacebookFragment extends Fragment {
 
         }
 
+
+
         @Override
         public void onCancel() {
