+++ JitsiForAndroid/app/src/main/res/layout/activity_contacts.xml
@@ -1,0 +1,185 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/main_layout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="fill_parent"
+        android:layout_height="50dp"
+        android:background="@drawable/lost_focus_style"
+        android:gravity="left">
+        <ImageView
+            android:id="@+id/image1"
+            android:layout_weight="0.2"
+            android:layout_width="0dip"
+            android:layout_height="wrap_content"
+            android:src="@mipmap/logo"/>
+        <TextView
+            android:layout_height="fill_parent"
+            android:layout_width="0dip"
+            android:layout_weight="0.8"
+            android:text="Title"
+            android:gravity="center_vertical"
+            android:textSize="30sp"
+            android:textColor="#000000"/>
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="fill_parent"
+        android:layout_height="50dp"
+        android:background="@drawable/lost_focus_style"
+        android:gravity="left">
+        <ImageView
+            android:id="@+id/image2"
+            android:layout_weight="0.2"
+            android:layout_width="0dip"
+            android:layout_height="wrap_content"
+            android:src="@mipmap/logo"/>
+        <TextView
+            android:layout_height="fill_parent"
+            android:layout_width="0dip"
+            android:layout_weight="0.8"
+            android:text="Title"
+            android:gravity="center_vertical"
+            android:textSize="30sp"
+            android:textColor="#000000"/>
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="fill_parent"
+        android:layout_height="50dp"
+        android:background="@drawable/lost_focus_style"
+        android:gravity="left">
+        <ImageView
+            android:id="@+id/image3"
+            android:layout_weight="0.2"
+            android:layout_width="0dip"
+            android:layout_height="wrap_content"
+            android:src="@mipmap/logo"/>
+        <TextView
+            android:layout_height="fill_parent"
+            android:layout_width="0dip"
+            android:layout_weight="0.8"
+            android:text="Title"
+            android:gravity="center_vertical"
+            android:textSize="30sp"
+            android:textColor="#000000"/>
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="fill_parent"
+        android:layout_height="50dp"
+        android:background="@drawable/lost_focus_style"
+        android:gravity="left">
+        <ImageView
+            android:id="@+id/image4"
+            android:layout_weight="0.2"
+            android:layout_width="0dip"
+            android:layout_height="wrap_content"
+            android:src="@mipmap/logo"/>
+        <TextView
+            android:layout_height="fill_parent"
+            android:layout_width="0dip"
+            android:layout_weight="0.8"
+            android:text="Title"
+            android:gravity="center_vertical"
+            android:textSize="30sp"
+            android:textColor="#000000"/>
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="fill_parent"
+        android:layout_height="50dp"
+        android:background="@drawable/lost_focus_style"
+        android:gravity="left">
+        <ImageView
+            android:id="@+id/image5"
+            android:layout_weight="0.2"
+            android:layout_width="0dip"
+            android:layout_height="wrap_content"
+            android:src="@mipmap/logo"/>
+        <TextView
+            android:layout_height="fill_parent"
+            android:layout_width="0dip"
+            android:layout_weight="0.8"
+            android:text="Title"
+            android:gravity="center_vertical"
+            android:textSize="30sp"
+            android:textColor="#000000"/>
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="fill_parent"
+        android:layout_height="50dp"
+        android:background="@drawable/lost_focus_style"
+        android:gravity="left">
+        <ImageView
+            android:id="@+id/image6"
+            android:layout_weight="0.2"
+            android:layout_width="0dip"
+            android:layout_height="wrap_content"
+            android:src="@mipmap/logo"/>
+        <TextView
+            android:layout_height="fill_parent"
+            android:layout_width="0dip"
+            android:layout_weight="0.8"
+            android:text="Title"
+            android:gravity="center_vertical"
+            android:textSize="30sp"
+            android:textColor="#000000"/>
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="fill_parent"
+        android:layout_height="50dp"
+        android:background="@drawable/lost_focus_style"
+        android:gravity="left">
+        <ImageView
+            android:id="@+id/image7"
+            android:layout_weight="0.2"
+            android:layout_width="0dip"
+            android:layout_height="wrap_content"
+            android:src="@mipmap/logo"/>
+        <TextView
+            android:layout_height="fill_parent"
+            android:layout_width="0dip"
+            android:layout_weight="0.8"
+            android:text="Title"
+            android:gravity="center_vertical"
+            android:textSize="30sp"
+            android:textColor="#000000"/>
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="fill_parent"
+        android:layout_height="50dp"
+        android:background="@drawable/lost_focus_style"
+        android:gravity="left">
+        <ImageView
+            android:id="@+id/image8"
+            android:layout_weight="0.2"
+            android:layout_width="0dip"
+            android:layout_height="wrap_content"
+            android:src="@mipmap/logo"/>
+        <TextView
+            android:layout_height="fill_parent"
+            android:layout_width="0dip"
+            android:layout_weight="0.8"
+            android:text="Title"
+            android:gravity="center_vertical"
+            android:textSize="30sp"
+            android:textColor="#000000"/>
+    </LinearLayout>
+
+
+
+</LinearLayout>
