+++ JitsiForAndroid/app/src/main/res/layout/activity_welcome.xml
@@ -4,28 +4,20 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
-    <ImageView
-        android:layout_width="100dp"
-        android:layout_height="wrap_content"
-        android:id="@+id/textViewImage"
-        android:src="@mipmap/jitsiorglogo"
 
-        android:layout_alignParentTop="true"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"/>
 
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
         android:paddingRight="@dimen/activity_horizontal_margin"
-        android:paddingTop="@dimen/activity_vertical_margin"
+        android:paddingTop="70dp"
         android:paddingBottom="@dimen/activity_vertical_margin" tools:context=".MainActivity">
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceLarge"
-            android:text="@string/emptyString"
-            android:id="@+id/TVusername" />
+        <!--<TextView-->
+            <!--android:layout_width="wrap_content"-->
+            <!--android:layout_height="wrap_content"-->
+            <!--android:textAppearance="?android:attr/textAppearanceLarge"-->
+            <!--android:text="@string/emptyString"-->
+            <!--android:id="@+id/TVusername" />-->
 
 
         <EditText
@@ -35,6 +27,7 @@
             android:hint="Enter Room Name"
             android:layout_alignBottom="@+id/createRoom"
             android:layout_alignParentLeft="true"
+            android:layout_marginRight="4dp"
             android:layout_alignParentStart="true"
             android:layout_toStartOf="@+id/createRoom" />
         <Button
