+++ JitsiForAndroid/app/src/main/res/values/strings.xml
@@ -6,4 +6,7 @@
     <string name="title_activity_register">register</string>
     <string name="emptyString" />
     <string name="title_activity_room">Meet</string>
+    <string name="title_activity_contacts">contacts</string>
+    <string name="title_activity_signup">signup</string>
+    <string name="title_activity_sign_up">sign_up</string>
 </resources>
