+++ JitsiForAndroid/app/src/main/res/values/styles.xml
@@ -15,4 +15,10 @@
         <item name="android:shadowDy" >1</item>
         <item name="android:shadowRadius" >2</item>
     </style>
+    <style name="AppTheme.NoActionBar">
+        <item name="windowActionBar">false</item>
+        <item name="windowNoTitle">true</item>
+    </style>
+    <style name="AppTheme.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+    <style name="AppTheme.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
 </resources>
